refactor: create lib and use in bin
this allows folks to use this crate as a library and/or as a CLI. that's
nice because then folks can build their own binary that bakes getset
functionality right in.

a likely use-case for that is building your own binary `setup` that
bundles in a `getset.toml` and runs it automatically for users and then
also does other things outside the scope of `getset`.